editted

+++ lab6_1.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 int main() {
     int a  =  20, b = 30;
-    float c = 4, output;
+    double c = 4, output;
 
     output = a - b/c;
     cout << "output: " << output << "\n";
